services/check_git_repo/pyproject.toml
[tool.poetry]
name = "check-git-repo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "check_git_repo"}]

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
PyYAML = "^6.0"
click = "^8.1.3"
GitPython = "^3.1.31"
#checkgit = {path = ".", develop = true}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.4.2"
pytest-snapshot = "^0.9.0"

[tool.poetry.scripts]
checkgit = 'main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# When using pipenv and setuptool to package
# [build-system]
# requires = ["setuptools", "wheel"]
# build-backend = "setuptools.build_meta:__legacy__"